created a quest-map model and transferred some functions here (quest.service, )
