+++ src/app/shared/services/quest.service.ts
@@ -116,61 +116,6 @@ export class QuestService {
 		);
 	}
 
-	private getQuestMapDetails(data: String): any {
-		var lines: any[] = data.split("\n");
-		var i = 0;
-		var dataArr: any[] = [];
-		var exArr: any[] = [];
-		var minX: number = 100;
-		var maxX: number = 0;
-		var basisY: number = 25;
-
-		let questPoint: any;
-		for (i = 0; i < lines.length; i++) {
-			var lineData = lines[i].split(",");
-			let x: number = parseInt(lineData[2]);
-			let y: number = parseInt(lineData[3]);
-			minX = minX < x && basisY == y ? minX : x;
-			maxX = maxX > x && basisY == y ? maxX : x;
-
-			if (lineData[1] === "scatter") {
-				questPoint = {
-					questId: lineData[0],
-					type: lineData[1],
-					x: x,
-					y: y,
-				}
-			} else if (lineData[1] === "line") {
-				questPoint = {
-					type: lineData[1],
-					x: x,
-					y: y,
-					x1: parseInt(lineData[4]),
-					y1: parseInt(lineData[5]),
-				}
-			} else if (lineData[1] === "exclude") {
-				let exclude: any = {
-					type: lineData[1],
-					x: x,
-					y: y,
-					direction: lineData[4]
-				}
-				exArr.push(exclude);
-				continue;
-			}
-
-			dataArr.push(questPoint);
-		}
-
-		var questMapDetails: any = {
-			maxX: maxX,
-			minX: minX,
-			questPositions: dataArr,
-			exclude: exArr
-		}
-		return questMapDetails;
-	}
-
 	/**
 	 * Returns the section's array quests
 	 * @param section_id id of the section whose array of quests needed to be retrieved
@@ -239,127 +184,6 @@ export class QuestService {
 	 */
 	joinQuest(user_id, quest_id, section_id) {
 		const url = this.sectionUrl;
-	}
-
-	setQuestMapDataSet(data: String): any[] {
-		let questMapDetails = this.getQuestMapDetails(data);
-
-		let exclude: any[] = questMapDetails.exclude;
-		let questPositions = questMapDetails.questPositions;
-		let minX: number = questMapDetails.minX;
-		let maxX: number = questMapDetails.maxX;
-		let basisY: number = 25;
-		let dataset: any;
-		var datasets: any[] = [];
-
-		for (let questPosition of questPositions) {
-			if (questPosition.type === "scatter") {
-				dataset = {
-					type: "scatter",
-					label: questPosition.questId,
-					data: [{
-						x: questPosition.x,
-						y: questPosition.y
-					}],
-					showLine: false
-				};
-				var scatterPoints: any[] = this.addQuestPlus(questPosition.x, questPosition.y, minX, maxX, basisY, exclude);
-				
-				if(scatterPoints.length != 0){
-					for(let scatterPoint of scatterPoints){
-						datasets.push(scatterPoint);
-					}
-				}
-			} else if (questPosition.type === "line") {
-				dataset = {
-					type: 'line',
-					label: 'Line Dataset 1',
-					data: [{
-						x: questPosition.x,
-						y: questPosition.y
-					}, {
-						x: questPosition.x1,
-						y: questPosition.y1
-					}],
-					fill: false,
-					radius: 0
-				}
-			}
-
-			datasets.push(dataset);
-
-		}
-
-		dataset = {
-			type: 'line',
-			label: 'Line Dataset 1',
-			data: [{
-				x: minX,
-				y: basisY
-			}, {
-				x: maxX,
-				y: basisY
-			}],
-			fill: false,
-			radius: 0
-		}
-
-		datasets.push(dataset);
-		console.warn(datasets);
-		return datasets;
-	}
-
-	addQuestPlus(x: number, y: number, minX: number, maxX: number, basisY: number, exclude: any[]): any[] {
-		let scatterPoints: any[] = [];
-
-		let excludedPoints = exclude.filter(data =>
-			data.x == x && data.y == y
-		);
-		let directions = ["N", "S", "E", "W"];
-		for (let direction of directions) {
-			if (y == basisY && (direction === "W" || (x < maxX && direction == "E"))) {
-				continue;
-			} else if ((y > basisY && direction === "S") || (y < basisY && direction === "N")) {
-				continue;
-			} else if (excludedPoints.filter(data => data.x == x && data.y == y && data.direction == direction).length != 0) {
-				continue;
-			}
-			let x1: number = x;
-			let y1: number = y;
-			if (direction == "N" || direction == "S") {
-				y1 = direction === "N" ? y + 2 : y - 2;
-			} else {
-				x1 = direction === "E" ? x + 2 : x - 2;
-			}
-
-			let scatterPoint: any = {
-				type: "scatter",
-				label: "Add Quest",
-				data: [{
-					x: x1,
-					y: y1
-				}],
-				showLine: false,
-				pointStyle: "cross"
-			}
-
-			scatterPoints.push(scatterPoint);
-		}
-
-		console.log(scatterPoints);
-		return scatterPoints;
-	}
-
-	addNewQuestLine(x, y, direction){
-		switch(direction){
-			case "N":
-			break;
-			case "W":
-			break;
-			case "E":
-			
-			break;
-		}
 	}
 
 	/**
